Fix remaining issues
