+++ app/src/main/java/ch/epfl/sweng/SDP/utils/AnimUtils.java
@@ -1,12 +1,16 @@
 package ch.epfl.sweng.SDP.utils;
 
 import android.content.Context;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
+import android.widget.TextView;
 
+import ch.epfl.sweng.SDP.Activity;
 import ch.epfl.sweng.SDP.R;
 import ch.epfl.sweng.SDP.home.BounceInterpolator;
+import ch.epfl.sweng.SDP.home.HomeActivity;
 
 public class AnimUtils {
 
@@ -63,5 +67,28 @@ public class AnimUtils {
         final Animation press = AnimationUtils.loadAnimation(context, R.anim.press);
         press.setFillAfter(true);
         view.startAnimation(press);
+    }
+
+    /**
+     * Sets listener and animation for exit button.
+     */
+    public static void setExitListener(final View exitButton, final Activity activity) {
+        exitButton.setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View view, MotionEvent event) {
+                switch (event.getAction()) {
+                    case MotionEvent.ACTION_DOWN:
+                        pressButton(exitButton, activity);
+                        break;
+                    case MotionEvent.ACTION_UP:
+                        bounceButton(view, HomeActivity.MAIN_AMPLITUDE,
+                                HomeActivity.MAIN_FREQUENCY, activity);
+                        activity.launchActivity(HomeActivity.class);
+                        break;
+                    default:
+                }
+                return true;
+            }
+        });
     }
 }
